+++ src/main/java/com/github/department/repo/EmployeeRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface EmployeeRepo extends CrudRepository<Employee, Integer> {
+public interface EmployeeRepo extends CrudRepository<Employee, Long> {
     List<Employee> findByName(String name);
 
     List<Employee> findByDepartment(Department department);
